src/ns_web_crawler/pyproject.toml
[tool.poetry]
name = "ns_web_crawler"
version = "0.1.0"
description = ""
authors = ["Yuchen Liu <steny138@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.8"
Scrapy = "^2.4.1"
SQLAlchemy = "^1.3.22"
uuid = "^1.30"
google-api-python-client = "^1.12.8"
python-dotenv = "^0.15.0"
requests = "^2.25.1"
xmltodict = "^0.12.0"
python-dateutil = "^2.8.1"
pytz = "^2020.5"
APScheduler = "^3.6.3"
psycopg2 = "^2.8.6"
scrapyd = "^1.2.1"
scrapyd-client = "^1.1.0"
gunicorn = "^20.1.0"
scrapydweb = "^1.4.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
